+++ lib/capybara_accessible_selectors/selectors/status.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+Capybara.add_selector(:status) do
+  xpath do |locator, *|
+    XPath.descendant[XPath.attr(:role) == "status"][XPath.string.is(locator.to_s)]
+  end
+end